Optimize code
The ck_json() is used to check that the json is valid.
I call ck_json with all the strings that I'm going to send to the
browser (all the chart data).
If there are errors in the json structure then the browser will
fail but the messages can be difficult to interpret, much less
where the error is.
So I call ck_json to check each chart's data.
But I don't need the full json object, just a 'is valid or not'.
If it is valid then continue. If not valid then do the full parse
where I can get error messages.
do_json() is called once per event and currently charts.json file
is parsed every time.
It can be very big.
So save off the parsed json object to avoid parsing.