save new_dvals vector values for binary der evts
The tc_syscalls.lua and gen_div_pair.lua routines are heavily
used. So I replaced the lua rtns with c++ rtns.
The c++ rtns syscall_der_evt() and gen_div_der_evt() still use the
'copy the new values to a vector of strings' for later conversion
back to doubles.
This commit creates a new_dvals vector for the c++ rtns and the
double values are saved as doubles. This has an increased cost
of a vector of doubles in each sample.
Although the conversion to string and back to doubles is avoided
I can't really measure a performance change (it seems to be within
the noise). And it does use more memory. But I wanted to have the
increased precision of the values (all other things being a wash).