Show Windows debugging options
The 'last' setting of a Makefile variable is the one that matters.
But this change makes it easy to set the debug compile/lib
options.
Also note that if you run with the --debug option (--debug must be
first option after oppat.exe) like:
bin\oppat.exe --debug -r ....
then the debugger can be attached at runtime to the oppat.exe.
This is done in oppat.cpp.
If you compile with the debug options and use --debug then there
is a call to _CrtSetDbgFlag() which sets more debug options.
The _CrtSetDbgFlag() options in oppat.cpp will really, really
slow down oppat.exe if you've compiled with the debug options.