Allow for ck_json to optionally return json object
Currently ck_json() verifies that the json is valid and discards
the result, just returning 'valid' or printing an error message
and exiting.
Sometimes we just want to check if it is valid (using accept())
and we are never going to fully parse the object.
Sometimes we are going to parse the object anyway so why not skip
the 'accept()' step, do the parse and return the object.
This commit implements this idea.