Have spin.cpp create phase file
Currently I try to pull the phase info from the markers in the
ETL trace file (on windows).
It is simpler to just create a phase file like I do with geekbench.
The '-p phase_file_name' cmdline option to spin.cpp creates the
phase file. The -p option is only checked if you use the
-f input_file option to pass the work type/args option.