Save off fields for drop_if_str_contains action
There is already a charts action filter_regex oper. This will
let you select only event data matching a regex in a given field.
But for the ETW syscall events, I want to drop events (not add
them to the chart data) if the 'area' field contains 'Wait'.
So I extened the filter regex stuff to do a 'drop_if_str_contains'
test (which is much quicker than a regex.