Add more ETW charts for OS_view
The SysCall events are every useful but generate a lot of data.
Not worrying about amont of data right now.
Also added a sort of histogram feature:
 charts->options->SUM_TO_INTERVAL
This option is for the total lines in a chart and it accumulates
the y variable in each time bucket (the event's end timestamp is
used to determine the bucket).