Add SUM_TO_INTERVAL code
Currently if one wants to make a histogram with the tot_lines it
doesn't really work. Especially if you have numerator denominator
fields in the data.
With SUM_TO_INTERVAL code (in the charts.json charts->options) the
y_value is accumulated in the event's ending interval.
Like a histogram. It is really a frequency in each bucket.