Add more debug info
Working through 'why do some total lines have an all zero 'av.1'
line' issue.
Currently, for ETW files, for derived events, if one of the
events needed by a derfived has no records, you still get a line
for the 'no data' line (but it is all zero).
I haven't worked through this all the way. For perf the derived
event wouldn't get selected since the 'no data' event wouldn't
be in the file. But ETW has some events in the header section even
though they don't appear in the file.
Current workaround for this issue is for main.js to mark the
'av.1' line as 'av.1 no data'.
But I probably should not send the 'area' for the chart or, better
yet, figure out why it is left as '1' and PageFault (for the
ETW Hardfault and PageFault chart.