Handle pid=-1 for Zombie thread sched_switch event
When a thread is being terminated, the last sched_switch state
shows 'Z' and the 'comm' is ':-1' and pid/tid is -1/-1.
This currently shows up in the charts as an 'unknown' process
(as a greyed out area) in the cpu_busy chart.
The prev_comm and prev_pid in the event data still has the 'real'
comm and pid.
so, if the event pid shows -1 and the event is sched_switch then
get the real comm and pid from the sched_switch data.