Chg by_var_hsh to use uint32 in unordered map
Currently the by_var_hsh looks up a double value.
But this hash is for the 'by_var' which means it should be for an
index value. The values are stored as doubles but used as an
index. So lets just use an uint32 for the key in the unordered
map.
Also add a chart option TOT_LINE_LEGEND_WEIGHT_BY_X_BY_Y.
This option says to order the legend by the height and width of
each value in the totline.
This makes the legend order correspond to the first line in the
legend is the line with the most values.