Add TOT_LINE_LEGEND_WEIGHT_BY_X_BY_Y to etw charts
The ETW syscall charts need a new ordering for the legend.
The TOT_LINE_LEGEND_WEIGHT_BY_X_BY_Y chart option enables ordering
the legennd by the height*width of each line in each bucket of the
tot_lines. This lets the legend order correspond to the apparent
order is (from most to least) the 'line with the most values' is
the 'line which appears first'.
This required lots of changes in main.js.