Speedup build_chart_lines() with by_var_hsh
build_chart_lines() was taking a long time (10+ seconds) for some
charts with lots of data.
The main issue was that the call to get_by_var_idx() wasn't
passing the unordered_map by value so it was making a copy of it.
Also changed the by_var_hsh to use a uint32 key instead of a
double. This change should be okay since (as far as I know) the
'by var' is always an integer.
Also avoided some unnecessary work when the chart option
OVERLAPPING_RANGES_WITHIN_AREA is used.