Speed up build_shapes_json()
Building the json strings for the IVAL and PTS array is costly.
There are some improvements that can be made.
The drop_trailing_zeroes() rtn doesn't need to be called for
non-floating point values.
There are improvements that can be made based on the x0/x1 and
y0/y1 rules.
For line charts, y0==y1 so we don't need to convert y1 if y0==y1.
for rect charts (cpu_busy) lots of values are vertical lines so
x0==x1.
Also, it is possible that an x or y will be the same as the
previous x or y. So check for this too.
These changes have sped up the IVAL PTS string stuff by 2x for
the cpu_busy chart.