Add timing info and fix fe_idx being uninitialized
Add timing info to various routines for trying to improve
performance.
A previous commit moved the check for 'if the event is the event
we are looking for' up in the loop of routine fill_data_table().
This change skipped setting samples[i].fe_idx for prf data.
This showed up as  ival[FE_IDX] == -1 for the PCT_BUSY_BY_CPU
chart (where the 'not sched_switch' events show up above the cpu
occupancy bars).
The PCT_BUSY_BY_CPU chart (title 'cpu busy') chart will not know
the event type for any event except the sched_switch event.
This commit checks if fe_idx == -1 and if so, sets fe_idx and
then checks if the current event type is == desired evt type.