Change standalone web page logic
Try to reduce memory usage by zeroing out the compressed data
string after decompressing. This really isn't so much data freed.
The compression is about 10x so the uncompressed size is much
bigger.
Also since I changed the 'draw a chart' code to use a loop now the
popup progress info doesn't get updated. Wrap the long running
stuff in a setTimeOut(..., 0); to give the screen a chance to
update.