Optimize and more timing info
Add more timing info.
Change prf_sample_to_string() to avoid duplicating the snprintf()
line.
Part of the prf_parse_text() logic looks for which event is on the
current text line.
Currently the search is by the event list order of appearance in
the file.
We've already read the binary file so we know the frequency of the
events.
So sort the event lookup list by most frequent to least frequent.
This cuts about have the time from prf_parse_text(). Most of the
time now is the sort of the samples.