Try to reduce memory usage for standalone html
For the standalone html files, the chart data is compressed in a
hex64 string. The data has to be decompressed into a string (a
json string), then the json string has to be parsed and then the
json data used to generate the chart.
Currently first I decompress the data and store the string in an
array. After I parse the string I don't need it any more so now
I reset each of the saved string array entry to "".
For some of the sample_html_files such as win_mem_bw4_pcm.html,
this reduces the peak memory usage by half (to 1 GB).
For others it didn't seem to help much.
But you probably can't (well I can't) load win_mem_bw4_pcm.html
with the debugger open. I get an ooops out of memory by chrome.
So I have to close the debugger before loading the file.
After the file loads you can open the debugger and search (in
the console) for '__mem' to see how much memory you used (on
chrome at least).