Display which file a data point comes from.
Currently when you hover over a point in a chart, you don't get
a line indicating which file a data point is comes from.
Now display the file in the balloon (hover popup).
The tot_lines don't display the filename yet. They probably
should... I think it just be the same as the underlying lines
from which the tot_line is computed. At this poing I think all the
data for a chart comes from the same file (but it might be
different events).