Change fill_data_table() to track event count
For non-ETW events, we have a loop over all the samples when we
change a prf_obj.
Currently I just flag whether we have any data for a particular
event.
It could be helpful to check whether we've processed all the
events in a prf_obj (you can have multiple events per prf_obj).
In particular, for the case where we have a rare event in prf_obj
with lots of other events then we could check if we've processed
all the events for a given type and break out of the loop if so.
This commit tracks gets the event count for each event (in the
early loop over all events in the prf_obj) and then tracks how
many of the event we are looking for have been processed.
When we get a 'not got the event we are looking for' condition
then we check if we've hit the expected event count and break
out of the loop if so.