Reduce mem usage & speedup standalone html
Currently all the decompress() steps are done and then the
decompressed JSON strings are passed to the parse_chart_data()
function.
This requires storing the json strings for each chart in an array
and then passing each string in the array to parse_chart_data().
Combining the decompress and parse_chart_data into 1 step reduces
memory usage and goes faster.
So replace:
 arr[j] = decompress(hex_str[j]);
 parse_chart_data(arr[j]);
with
 parse_chart_data(decompress_str(hex_str[j]);