Fix issue with window resizing and some cleanup
Currently when a window is resized you can get the tot_line
variables added again.
For instance I might see av.read on the legend the first time the
chart is drawn.
Then resize the window and I get av.av.read and av.read in the
legend.
The issue is that I update gjson.chart_data[i].subcat_rng and the
tot_line logic sees the previous av.read not as a tot_line
variable and adds an av.av.read.
With this commit, at the start of can_shape(), I check the
vaiables in gjson.chart_data[i].subcat_rng to see if they are
tot_line variables and if they are, I pop() them. The tot_line
variables are appended to gjson.chart_data[i].subcat_rng so the
pop() is okay.
If you never resize the screen you won't see this issue.
But opening the debugger in the side panel resizes the screen.
I really should have a popup to show that a redraw of everything
due to a resize is underway.

I also quit putting (referencing) gjson.chart_data[i] in
gcanvas_args[i][2]. It just adds to confusion to reference the
same underlying data multiple ways.