Fix memory leak on resize() or chg pixels high
Currently if you change the window size or change the height of
the charts in pixels (via left-hand size menu at the bottom) then
the memory usage increases.
The bigger the chart data, the bigger the mem leak.
You can see the memory usage on chrome if you stared chrome with
the options:
  --enable-precise-memory-info
For geekbench linux haswell data, the memory usage is about 1.7 GB
and each resize uses another 200+ MBs.
I muddled around a little looking at causes.
I changed the get_xy() and xlate() routines to not return an array
but this didn't seem to matter.
The resize logic and the 'change charts height in pixels' logic
calls the can_shape() routine and doesn't just do a redraw.

The main cause of the mem leak was not removing some elements
(and their children) when I go back through can_shape().
I added a routine remove_divs(ele_nm) to delete node and kids.
And I call remove_divs() in several places now.

Now the memory usage is fairly constant between resize'ings and
changing pixel heights. It's not exactly the same but bounces
around the same number.

I also added code to force garbage collection on chrome.
You have to start chrome with --js-flags="--expose-gc" and then
there is a window.gc() routine present.
The routine gets called if 'g_do_gc = true'.