Script to combine dashboard PNGs into big PNG
The dashboard feature takes the bar and pie charts of the cpu
diagram and puts them in their own PNG. This PNG gets downloaded
to the backend when you do a 'zoom all charts' or on the initial
draw.
So if I run the 'zoom to each phase script' mk_pngs_for_phase.bat
then I have a *_dash00*.png for each phase.
This script reads the dowloaded json file (which has the dashboard
charts values and shape) and combines each phase's PNG into 1 big
PNG.
The script shows
 1) how to read the json file in python,
 2) get the values of a particular chart (the GIPS (billions
    instructions/sec) chart),
 3) get the min/max x/y pixels used for all the charts (from the
    json 'shape' field
 4) sort the phase dashboards by GIPS
 5) clip the PNGs to just show the charts
 6) combine the PNGs
 7) write the GIPS value above each chart
 8) save the resultant PNG
The python script requires the python modules visvis and imageio
to be installed.