Add follow_proc flag
Currently the checking to see how much a cpu is running the
follow_proc process is done in main.js.
This change adds fields to keep track of the follow_proc usage
in oppat.cpp and pass the a flag to main.js if a shape uses the
follow_proc.