Add follow_proc handling to oppat.cpp
Currently the code for detecting whether the follow_proc is
running is done in main.js.
This can be done quicker in oppat.cpp.
This commit detects whether the follow_proc process is running on
a cpu and, if so, when the shape is generated (for passing to
main.js) a flag is added to the shape type to indicate that the
follow_proc is used by this 'shape'.
On ARM x64 linux (for my ubuntu build for raspberry pi 3 b+ at
least), when a process is launched by perf, the 'base module' can
show up as perf instead of say spin.x.
This messes up the follow_proc charts... it might show perf running
for several seconds instead of spin.x. If you look at the call
graph though, you can see that perf launched spin.x and then
spin.x is running.
So I added code to look through the the callgraph data for the
PCT_BUSY_BY_CPU chart to see if the follow_proc is in the
callstack.