Add dashboard chart fields
Currently most of the most important charts on the cpu_diagram are
kind of buried among all the other data.
This commit adds fields for a dashboard.
The cpu_diagram bar and pie charts are reproduced in the
dashboard. This is a kind of experimental feature.
The data and shape of each dashboard chart is included in the
json_data.json file sent back to the backend.