Add follow_proc, dashboard code
Currently the follow_proc logic is all done in main.js.
I add code to oppat.cpp to track the follow_proc cpu usage.
oppat.cpp checks the module running on each cpu and due to a
feature of ARM linux, I have to check the modules in the callstack
as well. ARM linux (at least for ubuntu ARM64 on raspberry pi)
will show the module as 'perf' and then the follow_proc (such as
spin.x) in the callstack (since perf starts spin.x).
Also I added a dashboard of sorts.
The (IMO) most important cpu_diagram charts (the pie & bar charts
shown on the cpu_diagram) are sort of obscured by the busyness of
the cpu_diagram.
So I wanted to show the bar/pie charts on a canvas by themselves.
I've added this feature. The data in the bar & pie charts as well
as the shape (in pixels) of the chart in the dashbard are sent to
the json_data.json file sent to the backend.
The dashboard gets converted to a PNG and sent to the backend
whenever a full redraw is done (such as by a zoom all). The
json_data.json is sent at the same time.
The idea is, for something like a geekbench run with 50 phases,
it can be hard to pick out things of interest.
Usually one would generate an S-curve showing some KPI sorted by
highest to lowest.
So I show something like this for the spin_on_ARM_A53 data.
I sort the dashboard PNGs by lowest to highest GIPS (giga/billion
instructions per second) and the clip/combine the dashboard PNGs
into 1 big PNG. see images\spin_on_ARM_A63_the_dashboard.png.
The combined PNG is generated by dash_vv.py.
Most of the code in this commit deals with generating the bar &
pie charts on dashboard in a certain order.
The '"dash":{"order":1,' dash->order field indicates the order
for drawing the bar/pie in the dashboard.