Show help info for dashboard charts
Duplicating the bar & pie charts on the dashboard resulted in the
bar/pie help info (shown in the element title popup) not working.
So get the pie/bar help working in both the dashboard and
cpu_diagram.
And add info showing from which chart the bar/pie is based (the
chart_tag, title and description).