Reduce tot_line divisions for --by_phase
Currently if I'm creating PNGs for each phase using the --by_phase
cmdline option, then the number of divisions for each tot_line is
set by the 'cpu_busy chart interval'*100.
For a long-running job like geekbench on haswell this can result
in 30,000 division for each chart with a tot_line. And if you have
something like syscall events with 100s of event on a chart then
this can blowup the memory usage. For gb on haswell on chrome
I sometimes run into the out-of-memory crash at 2.7 GBs as shown
by window.performance.memory totHeap.
The high res divisions is a holdover from when I did the
follow_proc stuff in main.js. Now I compute where the follow_proc
process is used in oppat.cpp and pass it to main.js.
So I don't need such a high res tot_line division now.
This change changes it to 500. This is still higher than the
default 100 divisions. But the memory usage is only 10% higher.
totHeap is about 980 MBs at 100 divisions and 1080 MBs at 500
divisions.
Also changed the when reset_OS_view_image_ready() is called.
Currently, when doing --by_phase and waiting for each phase to be
done, after the currenty PNGs are sent to the backend, I call
reset_OS_view_image_ready(). This clears the OS_view charts.
The place I call it currently causes the OS_view charts to be
cleared even after the phase is done.
We only need to clear the OS_view charts if we are going to draw
another phase.
This change only calls reset_OS_view_image_ready() before we get
ready to draw another phase.