Add dashboard PNG for geekbench on haswell
OPPAT can generate a movie for each phase... which can be useful.
The geekbench movie consists of the OS_view/cpu_diagram canvas for
each non-idle phase of geekbench.
I've added a dashboard canvas which pulls out the bar & pie charts
from the cpu_diagram.
I use the dash_vv.py to read the downloaded pat_dash0*.png files
and generate 1 png from all the 50 input PNGs.
So that's how I generate the picture.
What does it show?
Well, for one thing, that my method of mapping 'per cpu' data to
'per core' isn't perfect.
When HT is enabled and you want to map each HT result to the core
then how to do it?
Knowing which HT cpu maps to which core is the easy part.
Take the case where you have 'per core' cpu_diagram (such as you
can retire 4 UOPs/cycle per core).
When all the CPUs are busy the mapping of 'per cpu' events to a
core works pretty well.
But when a cpu is idle, things get strange. If only 1 of the cpus
on the core is busy that is simpler and easier to handle, but
if both cpus are sometimes busy, my main.js 'map_cpu_2_core' code
doesn't work very well.
For instance, looking at the dashboard, for 'single core Camera',
the Uops/cycle chart shows 5 uops/cyc on core 0 whereas the max
possible uops/cyc is 4.
Why does this happen?
Well we're computing a ratio 'rat uops retired' / 'cycles'.
If a core is mostly idle how do you map the values for each cpu
to the core. The event 'rat uops retired' is a 'per cpu' event.
The 'cycles' event is 'per core'.
By 'per cpu' and 'per core' I use the idea:
 if you add the 2 cpu's values (say X and Y) of an event together
 over the same interval, Should you get:
  1) X+Y then this is a 'per cpu' event
     For example, the 'rat uops retired' event is a per cpu event.
  2) 0.5*(X+Y) then the event is 'per core'
     For example, the cycles event is per core (by this definition).
     If the cpu runs at 2 GHz, of 1 sec I'd expect to see 2G cycles.
     Each CPU will show 2 billion cycles but I wouldn't say that
     the frequency of the system is 4 GHz over the 1 sec interval.
     The frequency is 2 GHz over the 1 sec interval.
     The 0.5*(X+Y) is a pretty stupid way to reduce the per cpu
     events to a core but this is what I do. It works well if all
     the CPUs are busy but can get noisy if only 1 cpu is busy on
     the core.
     I could get more clever about the mapping (probably) but it
     will always be an approximation (probably).
     For instance I could look how busy each cpu is over the
     interval and just use the intervals where 1 or both cpus
     is busy.... but I haven't done this yet.