Add 'cpu-clock time by cpu' chart
The perf event groups that I collect all use the cpu-clock
software as the trigger to collect the event group.
Currently though I don't actually plot the cpu-clock event.
This commit adds a plot for the cpu-clock event.
If the event is ~1.0 then the event (and the associated event
groups) is getting collected at the expected frequency by perf.
For the haswell geekbench charts, the chart shows that the value
is 1 +- 0.02.
This is kind of a 'sanity check' chart that what I think is
supposed to be happening (in terms of data collection) is actually
happening.