Add pinning, sync start and summing thread performance
Currently the memory bw tests threads are spawned and then they
start to work on the bw tests. This is okay on smaller chips but
for server chips (like with 256 threads) you want all the threads
to start at the same time so that the load on the system starts
at the same time and ends more or less at the same time.
The immediately starting threads can slow down the other threads
starting too. I saw the last started threads ending a few seconds
after the 1st started threads.
This change starts the threads one at a time and waits for them
to initialize before starting the next thread. The new thread
startup logic also (on linux) pins the code to a CPU to minimize
NUMA bad effects.
The threads then wait at a barrier until all the threads are up.
Code is also added to print out each threads performance in a
more readable manner.