More changes to compile on macbook. static link fails
More changes to avoid compile errors on macbook.
The zlib Makefile tries to do a static & dynamic build. The
dyn build isn't needed and gets link errors so I did:
cd zlib*; ./configure --static
to just build the .a lib (which is all I need).
These changes let oppat compile but the static link fails.
Apparently it is a std problem on OS X.
More work on this eventually.