Add support for measuring remote memory bandwidth
Currently for the mem_bw tests, a thread is started, pinned to
the cpu, and memory is allocated and accessed on that cpu.
So all memory is allocated locally and accessed locally.
On servers NUMA effects are important.
In the case where software isn't numa aware then memory might be
allocated on numa node X and accessed on some other node.
This remote access is slower than local access.
This change adds a simple remote memory bandwidth test.
Right now the test only deals with the first 2 nodes (even if
there are more than 2). And it only does prints the case where
you allocate on node 1 and read the memory from node 0.
A sample cmd line is
  bin/spin.x 10 mem_bw_remote 64 80m
which will run the remote bw test for 10 seconds with
a stride of 64 byte and an array size of 80m/cpu.
Remote bw is linux only at the moment.

This gives remote bw similar to the Intel memory latency checker
for the command line:
  mlc --bandwidth_matrix