Add 'freq' work type
On cloud boxes we can't access perf counters to get frequency
from cycles.
This commit adds a work type of 'freq' which does a spin cycle
of a known number of instructions. And added calls to get
cputime. With these 2 quantities we can compute frequency.
This assumes that the 'rorl' instruction will be done in 1 cycle.
Most of the instruction throughput articles I've seen indicate
this assumption is okay.