Add missing windows freq assembly and -c num_cpus
Currently the -w freq/freq_sml only works on linux/mac.
The inline assembly isn't supported on windows vs.
Create spin_wait/win_rorl.asm which does 10000 rorl eax,2
instructions. This seems to a 1 rorl/cycle instruction on Intel
and AMD.
So I can compute frequency with the instruction.
Also allowed setting the -l num_of_loops in the freq/freq_sml
logic so that we can better control how long a loop over the
assembly code takes. Eventually I hope to have a workload to
measure noisy neighbor on servers. And to set an RPS and get the
distribution of response times.

Also, currently the code assumes using all the cpus.
Doing the frequency work types, this gives you the 'what is the
frequency when all the cpus are busy'.
But the turbo frequency varies (on intel anyway) with the number
of busy cpus. When only 1 or 2 cpus are busy the freq is the max.
So added the -c num_cpus options to reduce the number of cpus.
I still pin the 1st thread to the cpu0, then next thread (if
-c 2+ or -c not entered and num_cpus >= 2) to cpu1 etc.