Change -c num_cpus to -n num_cpus, add -c clock option
It seemed more natural to use '-n number_of_cpus' and I wanted
to add a '-c t|s' to set the clock to use.
'-c t' uses the TSC for timing the loops for freq_sml.
'-c s' uses the OS time routines.
The default is the OS time routines.
The TSC is supposed to be invariant but there is a skew between
the CPUs and I can't prevent the thread migrating to a new cpu.
Plus it really seems that the TSC must be getting reset/bumped
changed by (I assume) the OS.
If you do:
./bin/spin.x -w freq_sml -c t -t 10 -v
then the variation in the TSC will be displayed.
It's not too bad but the difference varies run to run... which
makes it not too useful. I don't know where the variation is
coming from.