Add yield option for freq_sml work type
Currently freq_sml doesn't yield the cpu (voluntarily) until the
time limit is reached.
Since we can dial how many loops freq_sml does, we can dial in a
certain amount of work.
The purpose of this is to test the impact of running without
yielding vs doing a yield on server performance.