Increase the size of charts and the amount of data/event
Currently I've only used oppat on systems with maybe 4 cpus.
I've started using it on servers with 32+ cpus.
So more events, bigger data files.
This commit allows more records per event before the event is
dropped.
The new limit lets me do 10 seconds on a 32 cpu system with tons
of sched_switch context swtich events and the call stacks for
each event.
Also, increase the pixels high for each chart. The cpu_busy chart
has 32 cpus so it is pretty hard to read in 250 pixels high.
You can override this in the gui (with the left hand side menu).