Dramatically reduce data passed to browser
Previously I've only run oppat on systems with a handfull of cpus.
Now I'm running on servers with 32+ cpus.
Doing flamegraphs on 10 seconds of 32 cpus with lots context
switching generates lots of events.
This data was too much for the browser.
Most of the data was the call stacks and the text of the event
strings. For 2 seconds of running, the str_pool (string pool)
sent to the browser was 55 MBs. What I sent was okay for small
systems but doesn't work on larger boxes.
Previously the text strings were built in oppat.x and hashed and
then string was put in the str_pool and this str_pool was sent
to the browser. When it came time to display the event's info,
main.js would look up the text strings index in the str_pool and
grab that text string.
In redoing this, I saw that the txt string was built from lots
of other smaller strings (like process name, sched_switch event
text, line number in the perf output file, etc).
The line number made every text string unique. And the perf
sched_switch event strings are mostly duplicated too (over 100s
of thousands of switches). So almost everthing was being
duplicated countless times.
Now I replace the text string with an array of hash values for
each substring and, for the line number I use a negative number
to indicate that it should be replace with 'line ....'.
This all reduced the amount of data sent 36x (55 MBs -> 1.4 MB).
It also improves processing time.