Allow creating text strings from array of text hashes
Currently lots of text strings are passed to the browser.
Too many nearly duplicate strings... say maybe 100s of characters
with only the a line number being different.
This change handles replacing the text string with an array of
hashes of text substrings. This greatly reduces the amount of
data passes to the browser and reduces the memory used by main.js.