Add hash_escape_string routine
Currently the text from the call stacks is just read in and passed
in a json table to the browser. This can be a problem if one of
the strings contains something that is not valid json syntax.
For example, a string may contain " or other characters that need
to be escaped.
Since all the data in the callstack hash is going to be sent to
the browser in json, I added a hash_escape_string() routine.
This new routine first creates an valid json escaped version of
the string (using 2 nlohmann's MIT licensed routines). And if the
escaped string is not the same as the input string then the
escaped string (instead of the input string) is put into the hash.