Reduce memory usage on send & main.js
Currently the call stacks are converted to an array of indices
into the str_pool. That's good.
But the callstacks are duplicates, lots of duplicates.
So hash the string of integers and just pass the index to the
string of integers to main.js.
This reduces another 20% the memory usage. main.js will have to
convert the string of comma separated integers into an integer
array but that's no problem.
Also, change the txt_ai key for the 'text array indices' to 'taa'.
Since we are generating graphs with millions of points, every
byte helps (somewhat).