Handle hash of callstack and txt_ai->taa
Currently each callstack is passed as an array of indices into
the str_pool. That good but there is a lot of duplication in the
callstacks. So oppat.cpp hashes the comma separated list of
integer indices and just passes this new has into the str_pool.
This saves a bunch of memory (18% reduction in memory passed to
main.js for a 32 cpu, 10 sec job with a ton of context switches.
Also, since we are passing sooo many events, every byte not sent
saves memory for main.js. Change the json key 'txt_ai' to taa.
This saved another 6 MB.