Reduce memory usage by reduced myshapes.pts
For the cpu_busy chart, the y coordinates are a function of the
cpu number. There 2 main categories of things plotted:
1) the context switch event (a rectangular shape) and
2) a short vertical line above cswitch event showing another
event.
For 1), I only need to pass the x0 and x1 (beg and end time) of
the cswitch. The y0 and y1 can be computed from the cpu number.
For 2), I only need to pass the x0 of the event. x1 is == x0 and
y0 and y1 can be computed from the cpu number.
This reduced the memory usage sent to main.js by another 20%.
But chrome still reports lots of memory used. I'm not getting
an out-of-memory condition but I want to be able to send a lot
more data and improve the responsiveness. The zoom on the
cpu_busy can take maybe 10 seconds (it rebuilds the flamegraphs
to match the zoom interval). the Flamegraph build (for the 2
flamegraphs takes aobut 7 of the 10 seconds.