Add charts.json do_flamegraphs and flamegraph by comm/pid/tid
Currently the only way to not show flamegraphs (assuming there is
data for flamegraphs) is to change main.js g_do_flamegraphs=0.
This is sort of a debugging feature but it also reduces memory
usage in the browser and speeds up drawing the charts.
For instance, zooming the cpu_busy chart also zooms the
flamegraphs to the same interval and I've seen it take 10 secs
to zoom on a 32 cpu, tons of events data collection.
So add "do_flamegraphs":"y" or "n" to charts.json
This commit adds the field to the charts_default struct.
Also, currently the flamegraphs are drawn by default with a
'flame' for each comm+pid+tid. This is a lot of flames on a
server. Most of the time flamegraphs are just for 1 flame for
each comm (like java).
In the gui you can (in the lhs menu) select comm or comm+pid or
comm+pid+tid.
The default is comm+pid+tid.
This commit lets you set the default in charts.json.
"flamegraph_by_comm_pid_tid_default":"comm", will select comm.