Add support for flamegraph options and try to fix resize mem
Added support for disabling flamegraphs via the do_flamegraphs=y/n
option in charts.json. Also add flamegraphs_comm_pid_tid_default
option. This commit read the values from oppoat.cpp and uses the
values in drawing the charts.
I also noticed that when I do a resize of the 'pixels high' in
the lhs_menu then there is a memory leak.
I haven't figured out what is causing the leak yet. The workaround
is to set the default pixels_high in charts.json and not change it
in the gui.