Fix IO routines
The IO routines are not writing as much data as they are supposed
to write. Currently the input size is not used correctly to figure
out how much to read or write.
All that needs to be allocated is the IO buffer, not the whole
array (since there may not be enough memory anyway).
The -s size_str can have the suffix 'k' (for KBytes) or m (MBs) or
g (GBs).
Also fixed where the reads don't check for end of file.