Add signal handling, early exiting loops and GB array size
Currently spin.x runs until the -t time_limit is hit.
This commit is adds capturing signals (like killall -2 spin.x) and
exiting the loops when the signal is received.
Also, if the array size is a GGB or more (like -s 1g ) then the
array logic didn't work correctly. This commit fixes that.