Use -l loops option for mem_bw tests
Currently the -l number_of_loops option for the memory bandwidth
tests.
The loop count is adjusted automatically to reduce the number of
calls to the 'get time' routines.
The -t time is used to limit the seconds looping over the arrays.
Sometimes I want to do a fixed amount of work. This commit changes
the code to use the -l loops option if it is specified on the
command line. If -l loops is not specified then the code behaves
as before.