Add ping_pong and freq_fctr feature for freq_sml
To help debug perf events on Intel and AMD, I would like to be
able to control whether the other hyperthread of a core is busy
or not.
When I run -w freq_sml on both logical cpus of a core then both
threads are kept 100% busy.
To debug HT events, I added a --ping_pong run_secs,sleep_secs
With ping_pong, for every pair of cpus in the cpu_list, the 1st
cpu runs for run_secs while the 2nd cpu sleeps for sleep_secs.
Then the 2 cpus switch roles (the 1st cpu sleeps for sleep_secs
and the 2nd cpu runs for run_secs). Only 1 cpu of the pair of
cpus will run.
A cmd line would be
./spin.x -w freq_sml -t 10 -L 4,36 --ping_pong 0.5,0.5
I also added a freq_fctr option to let me set how much each cpu
is kept busy.
A cmd line would be
./spin.x -w freq_sml -t 10 -L 4,36 --freq_fctr 1.0,0.5
So the 1st cpu is kept busy 100% and the 2nd cpu is kept busy
50% of the time.
You can use the -l loops option to reduce the number of loops
in the freq_sml loop... this gives a finer grain of how much time
is done per iteration.