read /dev/cpu/*/msr to get cycles for freq_sml
Currently, when we do freq_sml (and related) work types, we
estimate the number of cycles based on the number of
'1 instruction per cyle' instructions we can do. This assumes
that we can actually do some instructions in 1 cycle.
On previous chips this worked okay.
On ice lake however, when we run freq_sml on both threads, we
can no longer get 1 instruction/cycle. We get an IPC = ~.91.
I've tried lots of different 1 cycle instructions and I can't find
any that can get IPC=1 if both HT threads are running.
So, on linux, see if /dev/cpu/*/msr is available.
You have to do: modprobe msr
This lets me report accurate frequencies on ice lake.
If I use /dev/cpu/*/msr then I print a message to that effect at
the bottom of the spin.x output.