modify --freq_fctr behavior for HT turbo studies
Currently you have to explicitly set the load factor for each cpu.
If there aare more cpu threads started than load factors then the
load gets set to 1 for the not-specified threads.
Change this behavior so that the last explictly entered load
factor is propagated to the rest of the threads.
Also print out more debug info (such as the frequency for each
thread). Currently the Gops/sec is used as the freq but, since it
uses elapsed time as the denominator in the freq equation then
if the load is not 1 then the freq is off.
This commit prints out a freq_fctr freq using cputime as the
denominator.
Also an average sleep time and an average runtime is printed out.